rest

+++ Laura/src/demo/PersRepository.java
@@ -35,7 +35,7 @@ public class PersRepository {
       criteria.select(member).orderBy(cb.asc(member.get("nachname")));
       return em.createQuery(criteria).getResultList();
   }
-  public PersEntity findById(Long id) {
+  public PersEntity findById(Integer id) {
       return em.find(PersEntity.class, id);
   }
 }
